+++ src/components/FacilitiesMap.js
@@ -1,16 +1,14 @@
 import React from 'react';
-import ReactMapboxGl, {Marker} from 'react-mapbox-gl';
-import {withStyles} from 'material-ui/styles';
-import {getMaxBounds} from '../utils/mapboxUtils';
-import MapDialog from './MapDialog';
+import ReactMapboxGl, {Marker, Popup} from 'react-mapbox-gl';
+import {MenuItem} from 'material-ui/Menu';
+import Select from 'material-ui/Select';
+import {FormControl} from 'material-ui/Form';
+import {getMaxBounds, getCenterOfCampusRegion} from '../utils/mapboxUtils';
+import mapboxgl from 'mapbox-gl';
+import {Typography} from 'material-ui';
+import {removeBrackets} from '../utils/nameUtils';
 
 const mapboxToken = 'pk.eyJ1IjoibWR1ZmZ5OCIsImEiOiJjaXk2a2lxODQwMDdyMnZzYTdyb3M4ZTloIn0.mSocl7zUnZBO6-CV9cvmnA';
-
-const Map = ReactMapboxGl({
-    accessToken: mapboxToken,
-    interactive: false,
-    attributionControl: false
-});
 
 const Mark = {
     backgroundColor: '#e74c3c',
@@ -20,104 +18,186 @@ const Mark = {
     border: '3px solid #EAA29B'
 };
 
-
 class FacilitiesMap extends React.Component {
     constructor(props) {
         super(props);
-        const maxBounds = getMaxBounds(props.campusRegion);
-        // const southWestBounds = [-77.321649, 38.823919]; //Coordinates for the south-west bound
-        // const northEastBounds = [-77.295213, 38.835720]; //Coordinates for the north-east bound
-        const southWestBounds = maxBounds[0]; //Coordinates for the south-west bound
-        const northEastBounds = maxBounds[1]; //Coordinates for the north-east bound
+        const {facility, interactive = true} = this.props;
+        const campusRegion = facility && facility.facility_location ? facility.facility_location.campus_region : 'fairfax';
 
+        this.Map = ReactMapboxGl({
+            accessToken: mapboxToken,
+            interactive: interactive,
+            attributionControl: false
+        });
+
+        const facilityLocationExists = facility && facility.facility_location && facility.facility_location.campus_region === campusRegion;
+
+        /**
+         * facilityLocations is an array of the type:
+         * {
+         *  location: {}
+         *  facilities: [{}, {}, ...]
+         * }
+         */
         this.state = {
-            positionReady: false,
-            position: {
-                longitude: 0,
-                latitude: 0
-            },
-            mappedRoute: false,
-            fitBounds: [southWestBounds, northEastBounds],
-            maxBounds: maxBounds,
+            maxBounds: getMaxBounds(campusRegion),
+            campusRegion: campusRegion,
+            center: facilityLocationExists ? facility.facility_location.coordinate_location.coordinates : getCenterOfCampusRegion(campusRegion),
+            zoom: facilityLocationExists ? [17] : [0],
             fitBoundsOptions: {},
-            mapDialogOpen: false
+            facilityLocations: [],
+            selectedLocation: null
         };
     }
 
-    handleRequestClose = () => {
+    componentWillReceiveProps(nextProps) {
+        const {facility, facilities} = nextProps;
+        const campusRegion = facility && facility.facility_location ? facility.facility_location.campus_region : 'fairfax';
+
+        this.changeRegion(campusRegion, facility);
+
+        if (this.state.facilityLocations.length === 0) {
+            this.generateLocationArray(facilities);
+        }
+    }
+
+    changeRegion = (campusRegion, facility) => {
+        if (!facility) {
+            facility = this.props.facility;
+        }
+
+        const facilityLocationExists = facility && facility.facility_location && facility.facility_location.campus_region === campusRegion;
+
         this.setState({
-            mapDialogOpen: false
+            maxBounds: getMaxBounds(campusRegion),
+            campusRegion: campusRegion,
+            center: facilityLocationExists ? facility.facility_location.coordinate_location.coordinates : getCenterOfCampusRegion(campusRegion),
+            zoom: facilityLocationExists ? [17] : [0]
         });
     };
 
-    render() {
-        const {facilities, facility, classes} = this.props;
-        const {fitBounds, maxBounds, fitBoundsOptions, mapDialogOpen} = this.state;
+    generateLocationArray = (facilities) => {
+        const locations = [];
 
-        let center, zoom;
-        try {
-            center = facility.facility_location.coordinate_location.coordinates;
-            zoom = [17];
-        } catch (e) {
-            center = [(maxBounds[0][0] + maxBounds[1][0]) / 2, (maxBounds[0][1] + maxBounds[1][1]) / 2];
-            zoom = [17];
+        facilities.forEach((facility) => {
+            const location = locations.find((loc) => loc.location.id === facility.facility_location.id);
+            if (location) {
+                location.facilities.push(facility);
+            } else {
+                locations.push({
+                    location: facility.facility_location,
+                    facilities: [facility]
+                });
+            }
+        });
+
+        this.setState({
+            facilityLocations: locations
+        });
+    }
+
+    selectLocation = (location) => {
+        const {interactive = true} = this.props;
+        const oldSelectedLocation = this.state.selectedLocation;
+        const oldZoom = this.state.zoom;
+
+        if (!interactive) {
+            return;
         }
 
+        this.setState({
+            selectedLocation: oldSelectedLocation !== location ? location : null,
+            center: location && location.location.coordinate_location.coordinates,
+            zoom: oldSelectedLocation !== location ? [17] : oldZoom
+        });
+    }
+
+    render() {
+        const {interactive = true} = this.props;
+        const {maxBounds, fitBoundsOptions, facilityLocations, selectedLocation, center, zoom} = this.state;
+
         return (
-            <div className={classes.mapContainer}>
-                <Map
-                    animationOptions={{
-                        animate: false
-                    }}
-                    onClick={() => {
-                        this.setState({
-                            mapDialogOpen: true
-                        });
-                    }}
-                    style="mapbox://styles/mduffy8/cjbcdxi3v73hp2spiyhxbkjde"
-                    movingMethod={'easeTo'}
-                    containerStyle={{
-                        height: '200px',
-                        width: '380px',
-                        margin: '10px',
-                        borderRadius: '3px',
-                        cursor: 'pointer'
-                    }}
-                    center={center}
-                    fitBounds={fitBounds}
-                    fitBoundsOptions={fitBoundsOptions}
-                    zoom={zoom}
-                    maxBounds={maxBounds}>
-                    {(facilities.length > 0) && facilities.map((item) => {
-                        return (
-                            <Marker
-                                key={item.slug}
-                                coordinates={item.facility_location.coordinate_location.coordinates}
-                                anchor="bottom">
-                                <div style={Mark} />
-                            </Marker>
-                        );
-                    })}
-                </Map>
-                <MapDialog
-                    open={mapDialogOpen}
-                    facilities={facilities}
-                    maxBounds={maxBounds}
-                    zoom={zoom}
-                    center={center}
-                    onClose={this.handleRequestClose}
-                />
-            </div>
+            <this.Map
+                onStyleLoad={(map) => {
+                    if (interactive) {
+                        map.addControl(new mapboxgl.GeolocateControl({
+                            positionOptions: {
+                                enableHighAccuracy: true
+                            },
+                            trackUserLocation: true
+                        }));
+                    }
+                }}
+                animationOptions={{
+                    animate: true,
+                    duration: 1500
+                }}
+                style="mapbox://styles/mduffy8/cjbcdxi3v73hp2spiyhxbkjde"
+                movingMethod="easeTo"
+                containerStyle={{
+                    height: '100%',
+                    width: '100%',
+                    borderRadius: '5px',
+                    cursor: 'pointer'
+                }}
+                center={center}
+                fitBounds={maxBounds}
+                fitBoundsOptions={fitBoundsOptions}
+                zoom={zoom}
+                maxBounds={maxBounds}>
+
+                {interactive &&
+                (
+                    <FormControl className={'facilities-map-campus-select'}>
+                        <Select
+                            disableUnderline
+                            value={this.state.campusRegion}
+                            onChange={(e) => this.changeRegion(e.target.value)}>
+                            <MenuItem value={'fairfax'}>Fairfax</MenuItem>
+                            <MenuItem value={'arlington'}>Arlington</MenuItem>
+                            <MenuItem value={'prince william'}>SciTech</MenuItem>
+                            <MenuItem value={'front royal'}>Front Royal</MenuItem>
+                        </Select>
+                    </FormControl>
+                )}
+
+                {(facilityLocations.length > 0) && facilityLocations.map((item) => {
+                    const location = item.location;
+
+                    return (
+                        <Marker
+                            key={location.id}
+                            coordinates={location.coordinate_location.coordinates}
+                            anchor="bottom"
+                            onClick={() => this.selectLocation(item)}>
+                            <div style={Mark} />
+                        </Marker>
+                    );
+                })}
+
+                {selectedLocation && (
+                    <Popup coordinates={selectedLocation.location.coordinate_location.coordinates} anchor="top">
+                        <div>
+                            <Typography type="subheading" align={'center'}>
+                                {selectedLocation.location.building}
+                            </Typography>
+                        </div>
+                        <div>
+                            <ul className={'facilities-map-popup-list'}>
+                                {selectedLocation.facilities.map((facility) => {
+                                    return (
+                                        <li key={facility.slug}>
+                                            <Typography type="caption">{removeBrackets(facility.facility_name)}</Typography>
+                                        </li>
+                                    );
+                                })}
+                            </ul>
+                        </div>
+                    </Popup>
+                )}
+            </this.Map>
         );
     }
 }
 
-const styleSheet = {
-    mapContainer: {
-        transition: '250ms ease-in-out',
-        width: '100%'
-    }
-};
-
-
-export default withStyles(styleSheet)(FacilitiesMap);
+export default FacilitiesMap;