+++ src/components/MapDialog.js
@@ -1,15 +1,9 @@
 import React from 'react';
-import ReactMapboxGl, {Marker} from 'react-mapbox-gl';
-import mapboxgl from 'mapbox-gl';
 import PropTypes from 'prop-types';
 import Dialog from 'material-ui/Dialog';
-
-const mapboxToken = 'pk.eyJ1IjoibWR1ZmZ5OCIsImEiOiJjaXk2a2lxODQwMDdyMnZzYTdyb3M4ZTloIn0.mSocl7zUnZBO6-CV9cvmnA';
-
-const Map = ReactMapboxGl({
-    accessToken: mapboxToken,
-    attributionControl: false
-});
+import FacilitiesMap from './FacilitiesMap';
+import IconButton from 'material-ui/IconButton';
+import CloseIcon from 'material-ui-icons/Close';
 
 class MapDialog extends React.Component {
     handleRequestClose = () => {
@@ -17,48 +11,25 @@ class MapDialog extends React.Component {
     };
 
     render() {
-        const {facilities, zoom, center, maxBounds, ...other} = this.props;
+        const {facility, facilities, campusRegion, open, width, height, fullScreen = false} = this.props;
 
         return (
-            <Dialog onClose={this.handleRequestClose} {...other}>
-                <Map
-                    onStyleLoad={(map) => {
-                        map.addControl(new mapboxgl.GeolocateControl({
-                            positionOptions: {
-                                enableHighAccuracy: true
-                            },
-                            trackUserLocation: true
-                        }));
+            <Dialog onClose={this.handleRequestClose} open={open} fullScreen={fullScreen}>
+                <div style={{
+                    height: height || '100%',
+                    width: width || '100%'
+                }}>
+                    <IconButton className={'map-dialog-close-btn'} onClick={this.handleRequestClose}>
+                        <CloseIcon />
+                    </IconButton>
 
-                    }}
-                    onClick={() => {
-                        this.setState({
-                            mapDialogOpen: true
-                        });
-                    }}
-                    style="mapbox://styles/mduffy8/cjbcdxi3v73hp2spiyhxbkjde"
-                    movingMethod={'easeTo'}
-                    containerStyle={{
-                        height: '500px',
-                        width: '600px',
-                        borderRadius: '5px'
-                    }}
-                    center={center}
-                    zoom={zoom}
-                    maxBounds={maxBounds}>
-                    {(facilities.length > 0) && facilities.map((item) => {
-                        return (
-                            <Marker
-                                key={item.slug}
-                                coordinates={item.facility_location.coordinate_location.coordinates}
-                                anchor="bottom">
-                                <img style={{
-                                    objectFit: 'contain'
-                                }} height={40} width={40} src={item.logo} />
-                            </Marker>
-                        );
-                    })}
-                </Map>
+                    <FacilitiesMap
+                        facilities={facilities}
+                        facility={facility}
+                        interactive={true}
+                        campusRegion={campusRegion}
+                    />
+                </div>
             </Dialog>
         );
     }