+++ src/index.js
@@ -11,8 +11,11 @@ import ReduxThunk from 'redux-thunk';
 import reducers from './reducers/index';
 import {MuiThemeProvider} from 'material-ui/styles';
 import theme from './theme';
-import './styles/build/whatsOpen.css';
+import './styles/whatsOpen.scss';
 import 'typeface-roboto';
+import '../public/manifest.json';
+import '../public/favicon.png';
+import '../public/apple-app-site-association';
 
 // Create a history of your choosing (we're using a browser history in this case)
 const history = createHistory();