+++ webpack/paths.js
@@ -1,0 +1,15 @@
+const path = require('path');
+const fs = require('fs');
+
+const appDirectory = fs.realpathSync(process.cwd());
+const resolveApp = (relativePath) => path.resolve(appDirectory, relativePath);
+
+module.exports = {
+    appRoot: resolveApp(''),
+    appHtml: resolveApp('public/index.html'),
+    appIndexJs: resolveApp('src/index.js'),
+    appBuild: resolveApp('build'),
+    appSrc: resolveApp('src'),
+    appOutputJs: 'static/js/bundle.js',
+    publicPath: ''
+};